+++ changelog.md
@@ -1,6 +1,7 @@
 # v1.2.0
 - made it so leveldata is now in a seperate file for each level instead of one big file to prevent lag when saving / loading
-- made loading and saving data asyncronous to reduce lag
+- added level cache to reduce unnecessary read operations when opening the menu popup
+- made loading the data asynchronous to prevent the game freezing while loading the menu popup
 - fixed an issue where 'remove pauses when AFK' would not remove pauses when exiting or resuming in certain ways
 # v1.1.0
 - Ported to 2.2081